# docs/swagger.yaml
basePath: /
definitions:
  controllers.GoalAndTasks:
    properties:
      goal:
        $ref: '#/definitions/models.Goal'
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
    type: object
  controllers.GoalResponse:
    properties:
      goal:
        $ref: '#/definitions/models.Goal'
    type: object
  controllers.ProgressRequest:
    properties:
      day:
        type: integer
      task_progress:
        items:
          $ref: '#/definitions/controllers.TaskAndStatus'
        type: array
    type: object
  controllers.ProgressResponse:
    properties:
      goal:
        $ref: '#/definitions/models.Goal'
      on_track:
        type: boolean
      total_point:
        type: integer
    type: object
  controllers.SuggestRequest:
    properties:
      goal:
        $ref: '#/definitions/models.Goal'
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
    type: object
  controllers.SuggestResponse:
    properties:
      plans:
        items:
          $ref: '#/definitions/models.SuggestedPlan'
        type: array
    type: object
  controllers.TaskAndStatus:
    properties:
      is_done:
        type: boolean
      task:
        $ref: '#/definitions/models.Task'
    type: object
  models.Goal:
    properties:
      name:
        type: string
      point:
        type: integer
      status:
        type: integer
    type: object
  models.PlanResponse:
    properties:
      day:
        type: integer
      tasks_today:
        items:
          $ref: '#/definitions/models.TaskResponse'
        type: array
    type: object
  models.SuggestedPlan:
    properties:
      day:
        type: integer
      plans_today:
        items:
          $ref: '#/definitions/models.TaskResponse'
        type: array
    type: object
  models.Task:
    properties:
      goal_id:
        type: integer
      name:
        type: string
      point:
        type: integer
    type: object
  models.TaskResponse:
    properties:
      name:
        type: string
      point:
        type: integer
    type: object
  utils.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  utils.SuccessResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: okane no kyouiku API
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      description: 'GETリクエストに対して{"message": "Hello, World!"}を返す'
      operationId: helloWorld
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Hello Worldのエンドポイント
  /api/v1/goals:
    post:
      consumes:
      - application/json
      description: Set a goal with associated tasks
      operationId: SetGoal
      parameters:
      - description: Goal and Tasks object
        in: body
        name: goal
        required: true
        schema:
          $ref: '#/definitions/controllers.GoalAndTasks'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Set a goal with tasks
      tags:
      - goals
  /api/v1/goals/progress:
    get:
      description: Get the goal details, accumulated points, and whether it's on track
      operationId: CheckProgress
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ProgressResponse'
      summary: Check the progress of a goal
      tags:
      - goals
  /api/v1/plans:
    get:
      consumes:
      - application/json
      description: Get a list of plans
      operationId: GetPlans
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PlanResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Get plans
      tags:
      - plans
  /api/v1/plans/suggested:
    get:
      description: Get a list of suggested plans
      operationId: GetSuggestedPlans
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PlanResponse'
            type: array
      summary: Get suggested plans
      tags:
      - plans
    put:
      consumes:
      - application/json
      description: Accept the suggested plans and update the status to "inprogress"
      operationId: AcceptSuggestedPlans
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Accept suggested plans
      tags:
      - plans
  /api/v2/goals:
    get:
      consumes:
      - application/json
      description: Get a list of goals
      operationId: GetGoal
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.GoalResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Get goals
      tags:
      - goals
  /api/v2/plans/suggest:
    post:
      consumes:
      - application/json
      description: ユーザーが設定した目標とタスクに基づいて日々のお手伝いプランを生成する
      operationId: suggestDailyPlans
      parameters:
      - description: 提案リクエストのボディ
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.SuggestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.SuggestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: 日々のお手伝いプランを生成するエンドポイント
      tags:
      - plans
  /api/v2/plans/today:
    get:
      consumes:
      - application/json
      description: ユーザーが指定した日のデイリープランを取得する
      operationId: GetTodayPlan
      parameters:
      - description: 取得する日の番号
        in: query
        name: day
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PlanResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: 指定された日のデイリープランを取得するエンドポイント
      tags:
      - plans
    post:
      consumes:
      - application/json
      description: Submit the progress of tasks for today's plan and store in the
        database
      operationId: SubmitTodayProgress
      parameters:
      - description: Progress request object
        in: body
        name: progress
        required: true
        schema:
          $ref: '#/definitions/controllers.ProgressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Submit progress for today's plan
      tags:
      - plans
swagger: "2.0"